# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_jackzampolin_shelf_internal_jobs.ProviderProgress:
    properties:
      completed:
        description: Completed during this job execution
        type: integer
      completedAtStart:
        description: Already completed before job started (from DefraDB)
        type: integer
      failed:
        description: Failed during this job execution
        type: integer
      queued:
        description: Currently queued/in-flight
        type: integer
      totalExpected:
        description: Total work units expected for this provider
        type: integer
    type: object
  github_com_jackzampolin_shelf_internal_jobs.RateLimiterStatus:
    properties:
      last_429_time:
        type: string
      rps:
        type: number
      time_until_token:
        $ref: '#/definitions/time.Duration'
      tokens_available:
        type: number
      total_consumed:
        type: integer
      total_waited:
        $ref: '#/definitions/time.Duration'
      utilization:
        type: number
    type: object
  github_com_jackzampolin_shelf_internal_jobs.Record:
    properties:
      _docID:
        type: string
      book_id:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      error:
        type: string
      heartbeat_at:
        type: string
      job_type:
        type: string
      last_progress_at:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.Status'
      status_reason:
        type: string
    type: object
  github_com_jackzampolin_shelf_internal_jobs.Status:
    enum:
    - queued
    - running
    - waiting_provider
    - completed
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusQueued
    - StatusRunning
    - StatusWaitingProvider
    - StatusCompleted
    - StatusFailed
    - StatusCancelled
  github_com_jackzampolin_shelf_internal_jobs.WorkerStatusInfo:
    properties:
      health:
        type: string
      in_flight:
        type: integer
      job_in_flight:
        type: integer
      job_parked:
        type: integer
      job_queued:
        type: integer
      parked_units:
        type: integer
      queue_depth:
        type: integer
      rate_limiter:
        $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.RateLimiterStatus'
      type:
        type: string
      workers:
        type: integer
    type: object
...
  internal_server_endpoints.CreateJobRequest:
    properties:
      job_type:
        type: string
      metadata:
        additionalProperties: {}
        type: object
    type: object
  internal_server_endpoints.CreateJobResponse:
    properties:
      id:
        type: string
    type: object
...
  internal_server_endpoints.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_server_endpoints.GetJobResponse:
    properties:
      _docID:
        type: string
      book_id:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      error:
        type: string
      heartbeat_at:
        type: string
      job_type:
        type: string
      last_progress_at:
        type: string
      live_status:
        additionalProperties:
          type: string
        description: Live status fields (only populated for running jobs)
        type: object
      metadata:
        additionalProperties: {}
        type: object
      pending_units:
        type: integer
      progress:
        additionalProperties:
          $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.ProviderProgress'
        type: object
      started_at:
        type: string
      status:
        $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.Status'
      status_reason:
        type: string
      worker_status:
        additionalProperties:
          $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.WorkerStatusInfo'
        type: object
    type: object
...
  internal_server_endpoints.JobStatusResponse:
    properties:
      book_id:
        type: string
      is_complete:
        type: boolean
      job_type:
        type: string
      metadata_complete:
        type: boolean
      ocr_complete:
        type: integer
      toc_extracted:
        type: boolean
      toc_found:
        type: boolean
      total_pages:
        type: integer
    type: object
...
  internal_server_endpoints.ListJobsResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.Record'
        type: array
    type: object
...
  internal_server_endpoints.UpdateJobRequest:
    properties:
      error:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      status:
        type: string
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/jackzampolin/shelf
  description: Book digitization pipeline API for managing books, jobs, and processing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Shelf API
  version: "1.0"
paths:
...
  /api/jobs:
    get:
      description: List all jobs with optional filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by job type
        in: query
        name: job_type
        type: string
      - description: Filter by book ID
        in: query
        name: book_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.ListJobsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: List jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Create a new job of the specified type
      parameters:
      - description: Job creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server_endpoints.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_server_endpoints.CreateJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Create a job
      tags:
      - jobs
  /api/jobs/{id}:
    delete:
      description: Delete a job by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Delete a job
      tags:
      - jobs
    get:
      description: Get detailed job information including live status for running
        jobs
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.GetJobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Get job by ID
      tags:
      - jobs
    patch:
      consumes:
      - application/json
      description: Update job status or metadata
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server_endpoints.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jackzampolin_shelf_internal_jobs.Record'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Update a job
      tags:
      - jobs
...
  /api/jobs/status/{book_id}:
    get:
      description: Get processing status for a book's jobs
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      - description: 'Job type (default: process-book)'
        in: query
        name: job_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.JobStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Get job status for a book
      tags:
      - jobs
...
swagger: "2.0"